+++ src/mem.rs
@@ -1,5 +1,7 @@
-pub use core::cell::UnsafeCell;
-pub use core::marker::{PhantomData, Unpin};
-pub use core::mem::{MaybeUninit, ManuallyDrop, size_of, size_of_val, align_of};
-pub use core::ptr::{NonNull, write_bytes, copy, copy_nonoverlapping};
+pub use core::{
+    cell::UnsafeCell,
+    marker::{PhantomData, Unpin},
+    mem::{MaybeUninit, ManuallyDrop, size_of, size_of_val, align_of},
+    ptr::{NonNull, write_bytes, copy, copy_nonoverlapping},
+};