+++ src/prelude.rs
@@ -1,0 +1,12 @@
+pub use crate::{
+    alloc::Alloc,
+    arena::Arena,
+    boks::Box,
+    hash::HashMap,
+    keyed::{Key, KSlice, KVec},
+    rc::Rc,
+    string::String,
+    vec::Vec,
+    traits::{FromIn, CopyIt, MapIt},
+};
+